Silence qemu_protect_none erorrs in M1 macs
This is a rather ugly workaround for the mprotect api returning error
codes when no permissions are given to a mapping on M1 macs. Given that
everything seems to still work fine and we don't know if this will get
fixed in the upstream darwin code, let's go with this workaround for the
time being.